Create Week 4- Database #1 Question 2.py
Write Python code to perform the following:
a) Create two tables.
b)Get the list of doctors as per the user inputted speciality and salary. Example: Pediatric doctors with salary above 30000.
c)Implement the functionality to fetch all the doctors as per the given Hospital Id. You must display the hospital name of a doctor.