Create Week 4- (Group) Database #3 Question 3.py
Question 3

Write a query that returns the HomeTeam, FTHG (number of home goals scored in a game) and FTAG (number of away goals scored in a game) from the Matches table. Only include data from the 2010 season and where ‘Aachen’ is the name of the home team. Return the results by the number of home goals scored in a game in descending order.

Print the total number of home games each team won during the 2016 season in descending order of number of home games from the Matches table.

Write a query that returns the first ten rows from the Unique_Teams table

Print the Match_ID and Unique_Team_ID with the corresponding Team_Name from the Unique_Teams and Teams_in_Matches tables. Use the WHERE statement first and then use the JOIN statement to get the same result.

Write a query that joins together the Unique_Teams data table and the Teams table, and returns the first 10 rows.

Write a query that shows the Unique_Team_ID and TeamName from the Unique_Teams table and AvgAgeHome, Season and ForeignPlayersHome from the Teams table. Only return the first five rows.

Write a query that shows the highest Match_ID for each team that ends in a “y” or a “r”. Along with the maximum Match_ID, display the Unique_Team_ID from the Teams_in_Matches table and the TeamName from the Unique_Teams table.